move MergeState, CompareStates and ParseRequiredStates to api package
